+++ mex/common/backend_api/connector.py
@@ -3,18 +3,17 @@ from urllib.parse import urljoin
 from requests.exceptions import HTTPError
 
 from mex.common.backend_api.models import (
-    ExtractedItemsRequest,
-    ExtractedItemsResponse,
     IdentifiersResponse,
-    MergedItemsResponse,
+    ItemsContainer,
     MergedModelTypeAdapter,
-    PreviewItemsResponse,
+    PaginatedItemsContainer,
     RuleSetResponseTypeAdapter,
 )
 from mex.common.connector import HTTPConnector
 from mex.common.models import (
     AnyExtractedModel,
     AnyMergedModel,
+    AnyPreviewModel,
     AnyRuleSetRequest,
     AnyRuleSetResponse,
 )
@@ -25,6 +24,7 @@ class BackendApiConnector(HTTPConnector):
     """Connector class to handle interaction with the Backend API."""
 
     API_VERSION = "v0"
+    INGEST_TIMEOUT = 30
 
     def _check_availability(self) -> None:
         """Send a GET request to verify the API is available."""
@@ -58,7 +58,8 @@ class BackendApiConnector(HTTPConnector):
         response = self.request(
             method="POST",
             endpoint="ingest",
-            payload=ExtractedItemsRequest(items=extracted_items),
+            payload=ItemsContainer[AnyExtractedModel](items=extracted_items),
+            timeout=self.INGEST_TIMEOUT,
         )
         return IdentifiersResponse.model_validate(response)
 
@@ -69,7 +70,7 @@ class BackendApiConnector(HTTPConnector):
         entity_type: list[str] | None,
         skip: int,
         limit: int,
-    ) -> ExtractedItemsResponse:
+    ) -> PaginatedItemsContainer[AnyExtractedModel]:
         """Fetch extracted items that match the given set of filters.
 
         Args:
@@ -96,7 +97,7 @@ class BackendApiConnector(HTTPConnector):
                 "limit": str(limit),
             },
         )
-        return ExtractedItemsResponse.model_validate(response)
+        return PaginatedItemsContainer[AnyExtractedModel].model_validate(response)
 
     def fetch_merged_items(
         self,
@@ -104,7 +105,7 @@ class BackendApiConnector(HTTPConnector):
         entity_type: list[str] | None,
         skip: int,
         limit: int,
-    ) -> MergedItemsResponse:
+    ) -> PaginatedItemsContainer[AnyMergedModel]:
         """Fetch merged items that match the given set of filters.
 
         Args:
@@ -129,7 +130,7 @@ class BackendApiConnector(HTTPConnector):
                 "limit": str(limit),
             },
         )
-        return MergedItemsResponse.model_validate(response)
+        return PaginatedItemsContainer[AnyMergedModel].model_validate(response)
 
     def get_merged_item(
         self,
@@ -155,7 +156,9 @@ class BackendApiConnector(HTTPConnector):
                 "limit": "1",
             },
         )
-        response_model = MergedItemsResponse.model_validate(response)
+        response_model = PaginatedItemsContainer[AnyMergedModel].model_validate(
+            response
+        )
         try:
             return response_model.items[0]
         except IndexError:
@@ -192,7 +195,7 @@ class BackendApiConnector(HTTPConnector):
         entity_type: list[str] | None,
         skip: int,
         limit: int,
-    ) -> PreviewItemsResponse:
+    ) -> PaginatedItemsContainer[AnyPreviewModel]:
         """Fetch merged item previews that match the given set of filters.
 
         Args:
@@ -207,7 +210,6 @@ class BackendApiConnector(HTTPConnector):
         Returns:
             One page of preview items and the total count that was matched
         """
-        # Note: this is forward-compat for MX-1649, backend might not support this yet!
         response = self.request(
             method="GET",
             endpoint="preview-item",
@@ -218,7 +220,7 @@ class BackendApiConnector(HTTPConnector):
                 "limit": str(limit),
             },
         )
-        return PreviewItemsResponse.model_validate(response)
+        return PaginatedItemsContainer[AnyPreviewModel].model_validate(response)
 
     def get_rule_set(
         self,